fix warning

+++ HomeBase/TCPServer.c
@@ -94,7 +94,7 @@ void tcpserver_run(void) {
 
         struct timeval tv;
         gettimeofday(&tv, NULL);
-        printf("You got a connection from client's ip %s, port %d at time %d.%d\n", inet_ntoa(client.sin_addr), htons(client.sin_port), tv.tv_sec, tv.tv_usec);
+        printf("You got a connection from client's ip %s, port %d at time %d.%d\n", inet_ntoa(client.sin_addr), htons(client.sin_port), (int)tv.tv_sec, (int)tv.tv_usec);
 
         int read_len;
         while(thread_running)
